backend/requirements.txt
uvicorn
fastapi 
pydantic
pydantic[email] 
pymongo 
motor 
bcrypt 
python-jose[cryptography] 
python-dotenv
python-multipart

